v1-basic-chatbot/pyproject.toml
[project]
name = "studybuddy"
version = "1.0.0"
description = "AI-powered study assistant - Chapter 1: Basic tutoring chatbot"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.0",
    "jupyter>=1.1.1",
    "uvicorn[standard]>=0.32.0",
    "openai>=1.54.0",
    "python-dotenv>=1.0.0",
]
